Rename index.ts to export.ts
