+++ packages/matter-node.js/tsconfig.dist.json
@@ -16,25 +16,25 @@
       "src/exports/cluster.ts",
       "src/exports/codec.ts",
       "src/exports/common.ts",
-      "src/crypto/index.ts",
+      "src/crypto/export.ts",
       "src/exports/datatype.ts",
       "src/exports/device.ts",
       "src/exports/fabric.ts",
       "src/exports/log.ts",
       "src/exports/math.ts",
       "src/exports/mdns.ts",
-      "src/net/index.ts",
+      "src/net/export.ts",
       "src/exports/protocol.ts",
       "src/exports/interaction.ts",
       "src/exports/securechannel.ts",
       "src/exports/schema.ts",
       "src/exports/session.ts",
       "src/exports/spec.ts",
-      "src/storage/index.ts",
-      "src/time/index.ts",
+      "src/storage/export.ts",
+      "src/time/export.ts",
       "src/exports/tlv.ts",
-      "src/util/index.ts",
-      "src/index.ts"
+      "src/util/export.ts",
+      "src/export.ts"
     ],
     "out": "../../docs/matter-node.js"
   }
