+++ packages/matter.js/src/fabric/export.ts
