+++ packages/matter.js/tsconfig.dist-cjs.json
@@ -12,30 +12,30 @@
   "include": ["src/**/*.ts"],
   "typedocOptions": {
     "entryPoints": [
-      "src/ble/index.ts",
-      "src/certificate/index.ts",
-      "src/cluster/index.ts",
-      "src/codec/index.ts",
-      "src/common/index.ts",
-      "src/crypto/index.ts",
-      "src/datatype/index.ts",
-      "src/device/index.ts",
-      "src/fabric/index.ts",
-      "src/log/index.ts",
-      "src/math/index.ts",
-      "src/mdns/index.ts",
-      "src/net/index.ts",
-      "src/protocol/index.ts",
-      "src/protocol/interaction/index.ts",
-      "src/protocol/securechannel/index.ts",
-      "src/schema/index.ts",
-      "src/session/index.ts",
-      "src/spec/index.ts",
-      "src/storage/index.ts",
-      "src/time/index.ts",
-      "src/tlv/index.ts",
-      "src/util/index.ts",
-      "src/index.ts"
+      "src/ble/export.ts",
+      "src/certificate/export.ts",
+      "src/cluster/export.ts",
+      "src/codec/export.ts",
+      "src/common/export.ts",
+      "src/crypto/export.ts",
+      "src/datatype/export.ts",
+      "src/device/export.ts",
+      "src/fabric/export.ts",
+      "src/log/export.ts",
+      "src/math/export.ts",
+      "src/mdns/export.ts",
+      "src/net/export.ts",
+      "src/protocol/export.ts",
+      "src/protocol/interaction/export.ts",
+      "src/protocol/securechannel/export.ts",
+      "src/schema/export.ts",
+      "src/session/export.ts",
+      "src/spec/export.ts",
+      "src/storage/export.ts",
+      "src/time/export.ts",
+      "src/tlv/export.ts",
+      "src/util/export.ts",
+      "src/export.ts"
     ],
     "out": "../../docs/matter.js"
   }
